+++ src/stapi_fastapi/routers/product_router.py
@@ -210,6 +210,7 @@ class ProductRouter(APIRouter):
                 ),
                 rel="create-order",
                 type=TYPE_JSON,
+                method="POST",
             ),
         ]