+++ tests/test_product.py
@@ -38,7 +38,9 @@ def test_product_response_self_link(
         url, body, "order-parameters", f"/products/{product_id}/order-parameters"
     )
     assert_link(url, body, "opportunities", f"/products/{product_id}/opportunities")
-    assert_link(url, body, "create-order", f"/products/{product_id}/orders")
+    assert_link(
+        url, body, "create-order", f"/products/{product_id}/orders", method="POST"
+    )
 
 
 @pytest.mark.parametrize("product_id", ["test-spotlight"])
@@ -100,6 +102,7 @@ def test_get_products_pagination(
                     "href": f"http://stapiserver/products/{product_id}/orders",
                     "rel": "create-order",
                     "type": "application/json",
+                    "method": "POST",
                 },
                 {
                     "href": f"http://stapiserver/products/{product_id}/opportunities",